+++ include/swift/AST/Stmt.h
@@ -20,6 +20,7 @@
 #include "swift/AST/Availability.h"
 #include "swift/AST/AvailabilitySpec.h"
 #include "swift/AST/ASTNode.h"
+#include "swift/AST/TypeAlignments.h"
 #include "swift/Basic/NullablePtr.h"
 
 namespace swift {
