+++ include/swift/AST/TypeAlignments.h
@@ -10,30 +10,43 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file defines the alignment of various Swift AST classes.
-//
-// It's useful to do this in a dedicated place to avoid recursive header
-// problems.
+/// \file This file defines the alignment of various Swift AST classes.
+///
+/// It's useful to do this in a dedicated place to avoid recursive header
+/// problems. To make sure we don't have any ODR violations, this header
+/// should be included in every header that defines one of the forward-
+/// declared types listed here.
 //
 //===----------------------------------------------------------------------===//
 
 #ifndef SWIFT_TYPEALIGNMENTS_H
 #define SWIFT_TYPEALIGNMENTS_H
 
+#include "llvm/Support/AlignOf.h"
+
 namespace swift {
   class AbstractStorageDecl;
+  class ArchetypeType;
+  class AssociatedTypeDecl;
+  class ASTContext;
+  class BraceStmt;
   class Decl;
   class DeclContext;
+  class Expr;
+  class GenericTypeParamDecl;
   class OperatorDecl;
   class ProtocolDecl;
   class ProtocolConformance;
-  class ArchetypeType;
+  class Stmt;
   class TypeVariableType;
   class TypeBase;
+  class ValueDecl;
 
   /// We frequently use three tag bits on all of these types.
   constexpr size_t DeclAlignInBits = 3;
   constexpr size_t DeclContextAlignInBits = 4;
+  constexpr size_t ExprAlignInBits = 3;
+  constexpr size_t StmtAlignInBits = 3;
   constexpr size_t TypeAlignInBits = 3;
 }
 
@@ -61,12 +74,25 @@ template <> class PointerLikeTypeTraits<CLASS*>           \
 
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::Decl, swift::DeclAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::AbstractStorageDecl, swift::DeclAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::AssociatedTypeDecl, swift::DeclAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::GenericTypeParamDecl, swift::DeclAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::OperatorDecl, swift::DeclAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::ProtocolDecl, swift::DeclAlignInBits)
-LLVM_DECLARE_TYPE_ALIGNMENT(swift::ProtocolConformance, swift::DeclAlignInBits)
-LLVM_DECLARE_TYPE_ALIGNMENT(swift::DeclContext, swift::DeclContextAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::ValueDecl, swift::DeclAlignInBits)
+
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::TypeBase, swift::TypeAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::ArchetypeType, swift::TypeAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::TypeVariableType, swift::TypeAlignInBits)
+
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::Stmt, swift::StmtAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::BraceStmt, swift::StmtAlignInBits)
+
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::ASTContext, 2);
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::DeclContext, swift::DeclContextAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::Expr, swift::ExprAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::ProtocolConformance, swift::DeclAlignInBits)
+
+static_assert(llvm::AlignOf<void*>::Alignment >= 2,
+              "pointer alignment is too small");
 
 #endif
