+++ include/swift/AST/Types.h
@@ -22,9 +22,9 @@
 #include "swift/AST/Ownership.h"
 #include "swift/AST/Requirement.h"
 #include "swift/AST/Type.h"
+#include "swift/AST/TypeAlignments.h"
 #include "swift/AST/Identifier.h"
 #include "swift/Basic/ArrayRefView.h"
-#include "swift/Basic/PointerLikeTypeTraitsFwdDecl.h"
 #include "swift/Basic/UUID.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/DenseMapInfo.h"
